refactor: move streamToAsyncIterable to private
